# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Language Learning API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  email?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  profileImageUrl?: string | null;
}

export interface GetCurrentAuthUserResponse {
  user: AuthUser | null;
}

export type LessonDifficulty =
  (typeof LessonDifficulty)[keyof typeof LessonDifficulty];

export const LessonDifficulty = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface Lesson {
  id: number;
  title: string;
  description: string;
  category: string;
  difficulty: LessonDifficulty;
  xpReward: number;
  wordCount: number;
  completed: boolean;
  locked: boolean;
  icon: string;
}

export interface Word {
  id: number;
  english: string;
  turkish: string;
  pronunciation: string;
  example: string;
  exampleTurkish: string;
}

export type QuestionType = (typeof QuestionType)[keyof typeof QuestionType];

export const QuestionType = {
  "translate-to-english": "translate-to-english",
  "translate-to-turkish": "translate-to-turkish",
  "multiple-choice": "multiple-choice",
  "fill-blank": "fill-blank",
  "listen-and-write": "listen-and-write",
} as const;

export interface Question {
  id: number;
  type: QuestionType;
  question: string;
  options?: string[];
  correctAnswer: string;
  hint?: string;
}

export interface LessonDetail {
  id: number;
  title: string;
  description: string;
  category: string;
  difficulty: string;
  xpReward: number;
  words: Word[];
  questions: Question[];
}

export interface UserProgress {
  xp: number;
  streak: number;
  level: number;
  completedLessons: number[];
  hearts: number;
  totalLessonsCompleted: number;
  dailyGoalXp: number;
  dailyXpEarned: number;
  longestStreak: number;
}

export interface CompleteLessonRequest {
  lessonId: number;
  score: number;
}

export interface SetDailyGoalRequest {
  goalXp: number;
}

export interface ResetResult {
  success: boolean;
  message: string;
}

export interface Achievement {
  id: string;
  title: string;
  description: string;
  icon: string;
  xpReward: number;
  unlocked: boolean;
  unlockedAt?: string | null;
  condition: string;
}

export interface NotebookWord {
  id: string;
  english: string;
  turkish: string;
  pronunciation?: string | null;
  example?: string | null;
  lessonId?: number | null;
  addedAt: string;
  reviewCount: number;
  known: boolean;
}

export interface AddToNotebookRequest {
  english: string;
  turkish: string;
  pronunciation?: string | null;
  example?: string | null;
  lessonId?: number | null;
}

export interface EmailRegisterRequest {
  email: string;
  password: string;
  name: string;
}

export interface EmailLoginRequest {
  email: string;
  password: string;
}

export interface ForgotPasswordRequest {
  email: string;
}

export interface ResetPasswordRequest {
  token: string;
  password: string;
}

export interface AuthActionResult {
  success: boolean;
  message: string;
}

export interface AuthError {
  error: string;
}

export interface LeaderboardEntry {
  rank: number;
  userId: string;
  displayName: string;
  xp: number;
  level: number;
  streak: number;
  profileImageUrl?: string | null;
  isCurrentUser: boolean;
}

export type VerifyEmailParams = {
  token: string;
};